Fix check for pinned shortcut in ShortcutRequestPinProcessor
Sometimes, when shortcut is removed, it's ID retained
in launcher's list of pinned shortcuts.
This makes it impossible to pin it in the future.
To fix it we check if we actually have the shortcut
before checking its pin status (like in requestPinShortcutLocked method).

Bug: 413285139
Test: Manual testing.

Change-Id: Ibb9693d6ebaa576595b3f2b15d18a8c73eb7fed9
(cherry picked from commit b02179df6d9d551b8a571d8c880234a8ea09f796)
Signed-off-by: Dmitrii <bankersenator@gmail.com>